+++ sdk/search/azure-search-documents/samples/async_samples/sample_buffered_sender_async.py
@@ -9,7 +9,7 @@
 """
 FILE: sample_batch_client_async.py
 DESCRIPTION:
-    This sample demonstrates how to upload, merge, or delete documents using SearchIndexDocumentBatchingClient.
+    This sample demonstrates how to upload, merge, or delete documents using SearchIndexingBufferedSender.
 USAGE:
     python sample_batch_client_async.py
 
@@ -27,7 +27,7 @@ index_name = os.getenv("AZURE_SEARCH_INDEX_NAME")
 key = os.getenv("AZURE_SEARCH_API_KEY")
 
 from azure.core.credentials import AzureKeyCredential
-from azure.search.documents.aio import SearchIndexDocumentBatchingClient
+from azure.search.documents.aio import SearchIndexingBufferedSender
 
 
 async def sample_batching_client():
@@ -39,16 +39,16 @@ async def sample_batching_client():
         'HotelName': 'Azure Inn',
     }
 
-    async with SearchIndexDocumentBatchingClient(
+    async with SearchIndexingBufferedSender(
             service_endpoint,
             index_name,
             AzureKeyCredential(key)) as batch_client:
         # add upload actions
-        await batch_client.add_upload_actions(documents=[DOCUMENT])
+        await batch_client.upload_documents(documents=[DOCUMENT])
         # add merge actions
-        await batch_client.add_merge_actions(documents=[{"HotelId": "1000", "Rating": 4.5}])
+        await batch_client.merge_documents(documents=[{"HotelId": "1000", "Rating": 4.5}])
         # add delete actions
-        await batch_client.add_delete_actions(documents=[{"HotelId": "1000"}])
+        await batch_client.delete_documents(documents=[{"HotelId": "1000"}])
 
 async def main():
     await sample_batching_client()