+++ sdk/search/azure-search-documents/tests/test_buffered_sender.py
@@ -7,8 +7,8 @@ try:
 except ImportError:
     import mock
 
-from azure.search.documents.aio import (
-    SearchIndexDocumentBatchingClient,
+from azure.search.documents import (
+    SearchIndexingBufferedSender,
 )
 from azure.core.credentials import AzureKeyCredential
 from azure.core.exceptions import HttpResponseError
@@ -16,49 +16,53 @@ from azure.search.documents.models import IndexingResult
 
 CREDENTIAL = AzureKeyCredential(key="test_api_key")
 
-class TestSearchBatchingClientAsync(object):
-    async def test_search_index_document_batching_client_kwargs(self):
-        async with SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, window=100) as client:
+class TestSearchBatchingClient(object):
+    def test_search_indexing_buffered_sender_kwargs(self):
+        with SearchIndexingBufferedSender("endpoint", "index name", CREDENTIAL, window=100) as client:
             assert client.batch_size == 100
             assert client._max_retry_count == 3
             assert client._auto_flush_interval == 60
             assert client._auto_flush
 
-
-    async def test_batch_queue(self):
-        async with SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, auto_flush=False) as client:
+    def test_batch_queue(self):
+        with SearchIndexingBufferedSender("endpoint", "index name", CREDENTIAL, auto_flush=False) as client:
             assert client._index_documents_batch
-            await client.add_upload_actions(["upload1"])
-            await client.add_delete_actions(["delete1", "delete2"])
-            await client.add_merge_actions(["merge1", "merge2", "merge3"])
-            await client.add_merge_or_upload_actions(["merge_or_upload1"])
+            client.upload_documents(["upload1"])
+            client.delete_documents(["delete1", "delete2"])
+            client.merge_documents(["merge1", "merge2", "merge3"])
+            client.merge_or_upload_documents(["merge_or_upload1"])
             assert len(client.actions) == 7
-            actions = await client._index_documents_batch.dequeue_actions()
+            actions = client._index_documents_batch.dequeue_actions()
             assert len(client.actions) == 0
-            await client._index_documents_batch.enqueue_actions(actions)
+            client._index_documents_batch.enqueue_actions(actions)
+            assert len(client.actions) == 7
+            actions = client._index_documents_batch.dequeue_actions()
+            assert len(client.actions) == 0
+            for action in actions:
+                client._index_documents_batch.enqueue_action(action)
             assert len(client.actions) == 7
 
 
     @mock.patch(
-        "azure.search.documents._internal.aio._search_index_document_batching_client_async.SearchIndexDocumentBatchingClient._process_if_needed"
+        "azure.search.documents._internal._search_indexing_buffered_sender.SearchIndexingBufferedSender._process_if_needed"
     )
-    async def test_process_if_needed(self, mock_process_if_needed):
-        async with SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL) as client:
-            await client.add_upload_actions(["upload1"])
-            await client.add_delete_actions(["delete1", "delete2"])
+    def test_process_if_needed(self, mock_process_if_needed):
+        with SearchIndexingBufferedSender("endpoint", "index name", CREDENTIAL) as client:
+            client.upload_documents(["upload1"])
+            client.delete_documents(["delete1", "delete2"])
         assert mock_process_if_needed.called
 
 
     @mock.patch(
-        "azure.search.documents._internal.aio._search_index_document_batching_client_async.SearchIndexDocumentBatchingClient._cleanup"
+        "azure.search.documents._internal._search_indexing_buffered_sender.SearchIndexingBufferedSender._cleanup"
     )
-    async def test_context_manager(self, mock_cleanup):
-        async with SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, auto_flush=False) as client:
-            await client.add_upload_actions(["upload1"])
-            await client.add_delete_actions(["delete1", "delete2"])
+    def test_context_manager(self, mock_cleanup):
+        with SearchIndexingBufferedSender("endpoint", "index name", CREDENTIAL, auto_flush=False) as client:
+            client.upload_documents(["upload1"])
+            client.delete_documents(["delete1", "delete2"])
         assert mock_cleanup.called
 
-    async def test_flush(self):
+    def test_flush(self):
         DOCUMENT = {
             'Category': 'Hotel',
             'HotelId': '1000',
@@ -66,21 +70,21 @@ class TestSearchBatchingClientAsync(object):
             'Rooms': [],
             'HotelName': 'Azure Inn',
         }
-        with mock.patch.object(SearchIndexDocumentBatchingClient, "_index_documents_actions", side_effect=HttpResponseError("Error")):
-            async with SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, auto_flush=False) as client:
+        with mock.patch.object(SearchIndexingBufferedSender, "_index_documents_actions", side_effect=HttpResponseError("Error")):
+            with SearchIndexingBufferedSender("endpoint", "index name", CREDENTIAL, auto_flush=False) as client:
                 client._index_key = "HotelId"
-                await client.add_upload_actions([DOCUMENT])
-                await client.flush()
+                client.upload_documents([DOCUMENT])
+                client.flush()
                 assert len(client.actions) == 0
 
-    async def test_callback_new(self):
+    def test_callback_new(self):
         on_new = mock.Mock()
-        async with SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, auto_flush=False, on_new=on_new) as client:
-            await client.add_upload_actions(["upload1"])
+        with SearchIndexingBufferedSender("endpoint", "index name", CREDENTIAL, auto_flush=False, on_new=on_new) as client:
+            client.upload_documents(["upload1"])
             assert on_new.called
 
-    async def test_callback_error(self):
-        async def mock_fail_index_documents(actions, timeout=86400):
+    def test_callback_error(self):
+        def mock_fail_index_documents(actions, timeout=86400):
             if len(actions) > 0:
                 print("There is something wrong")
                 result = IndexingResult()
@@ -91,19 +95,19 @@ class TestSearchBatchingClientAsync(object):
                 return [result]
 
         on_error = mock.Mock()
-        async with SearchIndexDocumentBatchingClient("endpoint",
+        with SearchIndexingBufferedSender("endpoint",
                                                "index name",
                                                CREDENTIAL,
                                                auto_flush=False,
                                                on_error=on_error) as client:
             client._index_documents_actions = mock_fail_index_documents
             client._index_key = "id"
-            await client.add_upload_actions({"id": 0})
-            await client.flush()
+            client.upload_documents({"id": 0})
+            client.flush()
             assert on_error.called
 
-    async def test_callback_progress(self):
-        async def mock_successful_index_documents(actions, timeout=86400):
+    def test_callback_progress(self):
+        def mock_successful_index_documents(actions, timeout=86400):
             if len(actions) > 0:
                 print("There is something wrong")
                 result = IndexingResult()
@@ -114,7 +118,7 @@ class TestSearchBatchingClientAsync(object):
 
         on_progress = mock.Mock()
         on_remove = mock.Mock()
-        async with SearchIndexDocumentBatchingClient("endpoint",
+        with SearchIndexingBufferedSender("endpoint",
                                                "index name",
                                                CREDENTIAL,
                                                auto_flush=False,
@@ -122,7 +126,7 @@ class TestSearchBatchingClientAsync(object):
                                                on_remove=on_remove) as client:
             client._index_documents_actions = mock_successful_index_documents
             client._index_key = "id"
-            await client.add_upload_actions({"id": 0})
-            await client.flush()
+            client.upload_documents({"id": 0})
+            client.flush()
             assert on_progress.called
             assert on_remove.called